Big "cleanup" moved code to separate files
